requirements.txt
cassandra-driver==3.29.2
confluent-kafka==2.5.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1